+++ app/data/courses/phat-hoc-pho-thong-quyen-3/khoa-thu-chin-luan-dai-thua-tram-phap-duy-thuc-va-nhon-minh-luan/bai-thu-5-mat-na-thuc.ts
@@ -346,7 +346,7 @@ const lesson: Lesson = {
   slug: 'bai-thu-5-mat-na-thuc',
   title: 'Bài Thứ 5 - Mạt na thức',
   type: 'article',
-  status: 'draft',
+  status: 'published',
   order: 7,
   createdAt: '2026-03-20',
   updatedAt: '2026-06-29',
