--- app/data/courses/phat-hoc-pho-thong-quyen-3/khoa-thu-chin-luan-dai-thua-tram-phap-duy-thuc-va-nhon-minh-luan/tap-nhut-luan-dai-thua-tram-phap-va-bat-thuc-qui-cu-tung.ts
@@ -1,15 +1,0 @@
-import type { Lesson } from '~/types/course'
-
-const lesson: Lesson = {
-  id: 'lesson-khoa-9-tap-nhut-luan-dai-thua-tram-phap-va-bat-thuc-qui-cu-tung',
-  slug: 'tap-nhut-luan-dai-thua-tram-phap-va-bat-thuc-qui-cu-tung',
-  title: 'Tập Nhứt: Luận Đại Thừa Trăm Pháp Và Bát Thức Qui Cũ Tụng',
-  type: 'article',
-  status: 'draft',
-  order: 2,
-  content: '<p>Nội dung bài học đang được biên soạn.</p>',
-  createdAt: '2026-03-20',
-  updatedAt: '2026-03-20',
-}
-
-export default lesson
